+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 2.12b2 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
+- Implement prefetch for Facebook.
+  [claytonc]
+
 - Add Open Graph fallback image for content lacking lead image;
   the image should be uploaded in the control panel configlet and must comply with best practices;
   canonical domain field was also moved to the new Open Graph field set in the configlet (closes `#130 <https://github.com/collective/sc.social.like/issues/130>`_).
@@ -16,9 +19,6 @@ There's a frood who really knows where his towel is.
 
 - Update Basque translation.
   [erral]
-
-- Implement prefetch for Facebook.
-  [claytonc]
 
 
 2.12b1 (2017-09-15)